Split code into separate files and reverted back to Jeff's original
