+++ app/hand.rb
@@ -1,0 +1,9 @@
+module Hand
+  def receive_card(card)
+    cards_held << card
+  end
+  
+  def clear_hand
+    self.cards_held = []
+  end
+end
